Add video poster logic to custom block edit component
